--- backend/routers/routes/users.js
@@ -1,8 +1,0 @@
-const express = require("express");
-const { createNewUser } = require("./../controllers/users");
-
-const usersRouter = express.Router();
-
-usersRouter.post("/", createNewUser);
-
-module.exports = usersRouter;
